Reverted weapon model render position to 0.
Weapon fire originates from the controller regardless of the position/shape/size/etc. of the weapon model.  I've modded the weapons mdls to match the fire origin, and so we need to revert the position of the models to the origin.  Any further moving needs to be to the MDLs, not in game.